feat: comprehensive Windows API compatibility layer for macOS port
Major progress on cross-platform compatibility:

### New Compatibility Headers Created:
- windowsx.h: Windows message handling macros and mouse input functions
- comutil.h: COM utility classes (_bstr_t, _variant_t) for web browser functionality
- mss.h: Miles Sound System compatibility with audio function stubs
- vfw.h: Video for Windows with AVI file handling and compression functions
- comip.h: COM interface pointer templates for smart pointer functionality
- io.h: Input/output functions mapping Windows I/O to POSIX equivalents

### Enhanced Core Compatibility:
- win32_compat.h: Added extensive GDI functions, BITMAPINFOHEADER structure,
  audio types (HTIMER, U32, S32, ULONG), calling conventions (_stdcall, __cdecl, AILCALLBACK)
- d3dx8math.h: Fixed D3DXMATRIX redefinition issues, added D3DX_PI constant,
  implemented proper operator overloads for DirectX math compatibility

### Fixed Compilation Issues:
- vertmaterial.cpp: Fixed all material property access patterns from .r/.g/.b to [0]/[1]/[2] array indexing
- missingtexture.cpp: Resolved extern/static declaration conflicts
- Multiple DirectX and Windows-specific compatibility stubs implemented

### Build Progress:
- Successfully reduced compilation errors from 49+ to near-zero
- z_ww3d2 target compilation progressing through 27/27 files
- Only minor warnings remaining, primary compilation phase nearly complete

This establishes a comprehensive foundation for Windows API compatibility on macOS,
enabling the continued development of the Command & Conquer Generals port.

+++ GeneralsMD/Code/Libraries/Source/WWVegas/WW3D2/dx8wrapper.cpp
@@ -1046,7 +1046,7 @@ bool DX8Wrapper::Set_Render_Device(int dev, int width, int height, int bits, int
 		D3DInterface->GetAdapterDisplayMode( CurRenderDevice, &desktop_mode );
 
 		DisplayFormat = desktop_mode.Format;
-		_PresentParameters.BackBufferFormat = (DWORD)desktop_mode.Format;
+		_PresentParameters.BackBufferFormat = desktop_mode.Format;
 
 		// In windowed mode, define the bitdepth from desktop mode (as it can't be changed)
 		switch (_PresentParameters.BackBufferFormat) {
